requirements.txt
python-dotenv==1.0.1
google-generativeai==0.6.0
gTTS==2.5.3
pyttsx3==2.90
Flask==2.3.2
flask-socketio==5.3.3
eventlet==0.33.3
